extract out zuul reporter
The reason we extract this is to be able to better intercept test events
and stack traces. We are able to sourcemap stacktrace errors and for
that need to be able to print our own traces over the default ones.

Allows for test reporting visualization to be the same across various ui
frameworks.