add `builder` config option
Allows user to specify a custom builder for zuul other than the default
browserify builder.

The builder should be a commonjs module that exports a single function:
function(files, cb). Files is an array of files and once the builder has
built them, it should call the callback with (err, src, map) where map
is an optional sourcemap object.

close #41