fix tape client finished state
Before this PR, the tape client would not output

1..9

At the end of tests when using phantombrowser for example, because we
closed the stream too early.

I am using tape-finished because that's the only easy way to know when a
tap suite is finished.

Otherwise we would have to parse 1..9, #tests 9 ourselves.

That's what was done before, in a module so we do not care.