Add support for config through js
We add support for configuration through JS by looking for a CommonJS module
named `zuul.config.js` for local configuration and treating `.zuulrc` as a
CommonJS module as well for global configuration (with fallback to YAML
parsing).

We exit if we encounter both `.zuul.yml` and `zuul.config.js` as porting from
one to another is straightforward and we don't want to introduce possible
confusion by merging them.

Closes #183.