Bump `engine.io` and `engine.io-client`
Closes vulnerability alerts, though they appear to have been false
positives in the context of `airtap`. The main motivation for
bumping these dependencies is simply keeping up with changes.

We're jumping several versions ahead but by using the transpiled
ES5 bundle of `engine.io-client` (included in its npm package) it
should not affect browser support of `airtap`. Verified with the
test matrix of `airtap-sauce` (which includes IE9).

Closes #276, closes #317, closes #312, closes #315.